Issue #16: Integrated Settings to SettingsAdapter.
